Restructured test to reduce code duplication.
